// apps/ui/aiwb/types/enums/models-table-fields.ts
// Copyright © Advanced Micro Devices, Inc., or its affiliates.
//
// SPDX-License-Identifier: MIT

export enum ModelsTableField {
  NAME = 'name',
  CANONICAL_NAME = 'canonicalName',
  CREATED_AT = 'createdAt',
  CREATED_BY = 'createdBy',
  ONBOARDING_STATUS = 'onboardingStatus',
  STATUS = 'status',
  WORKLOADS = 'workloads',
}